move cra to sub folder
