+++ create-react-app/public/manifest.json
